// src/logger/interfaces/logger-options.interface.ts
import { FastifyReply, FastifyRequest } from 'fastify';
import { DestinationStream, LevelWithSilent, LoggerOptions as PinoLoggerOptions } from 'pino';

export interface ILoggerOptions {
    pinoOptions?: PinoLoggerOptions;

    stream?: DestinationStream;

    customLogLevel?: (req: FastifyRequest, res: FastifyReply) => LevelWithSilent;

    ignorePaths?: string[];

    logFormat?: LogFormat;

    reqResSerializers?: {
        req: (req: FastifyRequest) => Record<string, any>;
        res: (res: FastifyReply) => Record<string, any>;
    };
}

export enum LogFormat {
    PRETTY = 'pretty',
    JSON = 'json',
}